Refactor WorkoutsController
Implement async in all methods in the controller, and take out the logic that is associated with the creation method
Create #createAll method in WorkoutExercises model that will take exercises and workout, then make and return all WorkoutExercises in an array.
Create #mapExercises that takes exercise names,  finds the exercises, and returns all in an array.